feat: implement (*cliz.Command).Run() func
